Add yarn workspace caching (second attempt) (#325)
## Description
This change is a retry of https://github.com/airplanedev/lib/pull/318 but without the improved caching behavior if there are any pre/post-install hooks in the user code or a provided airplane preinstall script. The problem with the latter is that they might assume that the entire code is present, which can cause the build to fail.

cc: @leeweisberger 

## Test plan
Testing completed successfully via adding a new unit tests to cover the hook behavior. I've verified that the test fails if I remove the hook checks but passes with them.