Move task info endpoint to internal (#706)
## Description
The editor now expects the task info to be under `/i/tasks/get` - we copy the implementation from `/v0/tasks/get` and delete the old handler. Also deletes a couple external endpoints that aren't used from the local dev server.

## Test plan
Tested loading a task page in the editor and verifying that it worked.