List remote resources alongside local resources (#707)
## Description

Return remote resources in the `/i/resources/list` endpoint of the local dev server if a fallback environment is specified. This will be used for listing remote resources in the editor sidebar, and is also required by the run details page for builtins executed against a remote resource.

## Test plan
Tested executing a task with a builtin and verified that run details page was populated with remote resource configuration.