Fix fallback behavior for views (#723)
## Description

Right now, if a fallback environment is set, a view will always execute it remotely, even if it exists locally. We should use the local client in views dev, which will handle logic to have a local task take precedence, and proxy the request to the remote environment if it's not available locally.

## Test plan
Tested locally with a fallback environment using a view with a task-backed component, where the task existed both locally and in the task backed environment.