Always fetch and store remote env (#730)
## Description
Always fetches a remote environment, even when no fallback environment is set. We fetch a few default remote resources, and in order for the user to know which environment we're pulling them from, we now store the remote environment (when no fallback environment is set, this is always the user's default environment). To determine whether we should use this remote env for fetching tasks/resources/etc. _besides_ of the default ones, we store a `UseFallbackEnv` variable in the local dev state.

## Test plan
This is mostly a refactor - verified that existing fallback environment behavior still works, and that the appropriate environment is included for default remote resources when listing resources even when no fallback environment is set.