Implement bundle based discovery (#345)
## Description
Implement a new discoverer that discovers bundles rather than tasks/views.

This discoverer is similar to the existing one, but it doesn't go so far as to get view/task configs. For each type of discoverer, it gets the root of the potential task/view.

Each bundle stores 
1. The path to the root of the bundle.
2. The target paths relative to the root that match entities that should be deployed.
3. (TODO) the build type and version of the bundle.

## Test plan
I added unit tests with fixtures. Nobody is using this yet.