Handle imports with node: scheme (#358)
## Description

The latest version of `js-sdk` introduces a dependency on `node-fetch` which uncovered issues with the workflow build process, in that it errored if code depended on imports that use a `node:` import scheme (e.g. `node:stream/web`).

This PR fixes that by aliasing `node:` imports the same way as if they were imported without the scheme.

Resolves TSK-926

## Test plan

Build our dogfooding workflows with the latest version of the JS SDK.