Determine node version from package.json (#354)
## Description
Determine the node version from the standard location in the package.json.

Make node version non-required in the task definition - this is the new recommended way to set it. We still support the old way for backwards compat purposes.

## Test plan
Unit tests for the version logic . Also more e2e unit tests for discovery.