Allow multiple tasks with the same root to run locally in parallel (#361)
## Description

There is currently a race condition where, if multiple tasks are being run in parallel, the last task run wipes out the state of the `.airplane` directory before the other tasks are finished running. This can happen, for example, if a view uses multiple task-backed components (e.g. many of our templates). This leads to error messages indicating that the shim is not found, etc.

To fix this, we do something similar to what we do for views, where instead of writing shim files to `.airplane` directly, we create them in a `.airplane/{task_slug}/` subdirectory. Ideally, we combine the `.airplane/` and `.airplane-view/` folders - however, the former is temporary and deleted after local task execution, while the latter is kept around to keep view development fast.

The only caller of `PrepareRun` is the local dev server, and so we can safely make this change as long as we have the corresponding CLI change.

## Test plan
Tested locally by ensuring node/python/shell tasks still ran locally, and also tested task composition with all of these (node -> node, node -> python, node->shell, node -> node -> node, etc.)