Airplane config (#366)
## Description
Create airplane config yaml. 

Currently the airplane config only contains the node version and env vars.

The node version is read in when determining the node version (superseded by package.json).

So far nobody reads env vars. We are going to read these during dev and in the server during build. I'll add this functionality as follow ups.

cc @calufornia example of adding a jsonschema which we want to do for the airplane.dev.yaml

This is undocumented, so this should be a no-op.

## Test plan
Unit tests