Add builtins to `lib` (#371)
## Description

Also adds GraphQL to the definition, which enables initializing + deploying a GraphQL task. The larger patch here is to accommodate builtin tasks generally.

Note that this does not generate a commented YAML file for builtins. Punting on this, because I think it might be harder than it's worth right now.

Resolves TSK-946.

## Test plan

Tested locally with updated CLI + API.