Make choice of slim vs. full base images explicit (#379)
## Description
This change updates our node and python task kind configs to add a `base` parameter that can be set to `"slim"` to indicate that we should use a slim base image when building. This change replaces the forced-slim behavior for workflows that was added in https://github.com/airplanedev/lib/pull/375.

The benefits of this change are that (1) non-workflow users can also use slim images and (2) workflow users can use non-slim images if they need to for whatever they're doing.

A few notes:

- I'm using a string instead of a bool so that we can possibly support other base image variants in the future (e.g., alpine vs. ubuntu)
- If base is unset, we use a non-slim image. Thus, the change should be a no-op for existing task configs. 
- Let me know if you think the term `base` isn't descriptive enough; we could use something like `baseImageType` instead?

cc: @colinking 

## Test plan
Testing completed successfully via extra unit tests and building locally with full and slim base configs.