Load sql entrypoint in runtime (#384)
Resolves TSK-1008

## Description
If someone changes a task's sql file without reloading the task config itself, we want to read the sql entrypoint and run the most up to date sql file.

## Test plan
Made the cli build with the change, tested studio doesnt have to reload the task but still runs the most up to date sql.