Don't attempt to open browser in sandbox (#784)
## Description

Adds the ability to bypass the "open browser" behavior when launching the studio, which will not be used by the remote sandbox, i.e. now we just start the local dev server and wait for a sigterm. If we don't do this, then an error occurs inside the sandbox container:
```
Started studio session at https://app.airplane.dev/studio?host=http://localhost:4000&__env=asdf (^C to quit)
Press ENTER to open the studio in the browser.

Something went wrong. Try running the command with the --debug flag for more details.
```

We gate this behind the `--sandbox` flag and hide it, since it should only be used by the remote studio.

Also removes the port from the local dev server state, which is unused.

## Test plan

Tested with and without `--sandbox` flag and verified behavior was as expected.