Expose local dev server in sandbox mode (#785)
## Description

The local dev server will be running inside the sandbox container for remote studio. In order to properly route requests to it, we bind the local dev server in this case to `0.0.0.0`, the all interfaces address. We don't do this for normal local dev because it prompts the user to accept incoming connections, and so we gate it behind the `--sandbox` flag: 
![image](https://user-images.githubusercontent.com/16730188/200596017-e62ac06a-f3a7-4a54-9e5f-cff289e3e84b.png)


## Test plan
Tested with and without flag, and verified that when `--sandbox` was supplied we could make requests to a local dev server running inside a docker container.