Don't error if task-specific subdirectory already exists (#397)
## Description
We'll occasionally see local task execution flakiness due to the `.airplane` directory not getting cleaned up directory, which also leaves behind the task-specific subdirectory in `./airplane/{task_slug}`. In the case that this subdir exists, we ignore the error when attempting to create it.

This also fixes task execution in the general case where an `.airplane/` directory failed to get cleaned up (although this directory should get destroyed/recreated on subsequent studio startups anyway).

## Test plan
Tested in conjunction with studio.