Support old tls version (#409)
## Description
When we upgraded to go1.18, we had to set a flag to support old TLS versions: https://github.com/airplanedev/airport/pull/2208

This flag is no longer supported in go1.19, so to upgrade to that we have to use TLS in our db drivers.

We seem to only use tls for mysql (does this make sense)? Here I'm setting a custom tls config that sets the min version of tls that we support to 1.0 from go1.19's default of 1.2

## Test plan
Have not tested this.