Don't fill in the default timeout value since it gets set in the API (#422)
## Description
We want allow workflows to have a timeout value of 0. Stop setting a default timeout to 3600 on deploy, since timeout validation and defaults are already handled in the API.

This requires a CLI change to ingore diffs where the task yaml doesn't have a timeout defined (but a previous version in the db has it filled out as 3600).

## Test plan
- [x] No timeout specified in task yaml still results in 3600 in the db (it gets filled in from the API)