Stop setting default timeout at deploy time (#800)
## Description
The API already sets a default timeout if it isn't defined.
This pulls in lib changes to stop doing that:
https://github.com/airplanedev/lib/pull/422/files

The change here is to NOT show a diff if the user doesn't have the timeout defined in their yaml. 
Previously, we were relying on setting the default at deploy time to be 3600 to calculate as no diff.

- [x] TODO: use published version of lib 
## Test plan
- [x] No change: deploying a task works. I deployed and ran the task. Same with a workflow.
- [x] Deploying a task with no timeout specified doesn't show the diff about timeout being different
- [x] Deploying a task where I change the timeout to something like 3000 (instead of the default 3600) does show a diff 
- [x] workflows are inline, which doesn't show diffs