Init inline JS tasks without a 0 timeout field (#811)
## Description
Uses this version of lib:
https://github.com/airplanedev/lib/pull/429

In https://github.com/airplanedev/lib/pull/422 we stop setting the timeout to 3600 at deploy time if it's not specified (aka 0 value for an int). The API is responsible for setting default timeouts to 3600. 

This fixes JS **inline** `tasks init` so it doesn't generate new inline JS tasks with a timeout of 0, instead just not having the field by default.

## Test plan
- [x] `airplane tasks init --inline` for JS has no workflow field
- [x] `airplane tasks init --inline --from slug` doesn't show a workflow field if it's 0 or 3600