Error if task has an invalid config attachment (#809)
## Description

Previously, if we attached an invalid config to a task, this would be silently ignored and usage of the config would just result in undefined (e.g. if referencing `process.env.MY_CONFIG`). Now, we validate that the config is included as part of the airplane.dev.yaml file, and error at execution time if not. Because we want to show this error to the user during the run, we move some materialization logic into the `Execute` function when preparing the run.

## Test plan
Valid config attachments still work.

Env var from config:
![image](https://user-images.githubusercontent.com/16730188/204368202-5b9b4133-469e-4f98-96cd-e4136b043e15.png)

Config attachment: 
![image](https://user-images.githubusercontent.com/16730188/204368267-fdbf8bb4-d0b5-4e6b-b533-7798f3c9d03d.png)