Deterministically generate dev resource id (#817)
## Description

A follow up to https://github.com/airplanedev/cli/pull/804 — dev resource IDs must be deterministic, because we generate resource IDs when loading in the dev config file. Previously, when dev resource IDs were random, this led to unintended behavior when creating a resource and deleting it immediately, deleting multiple resources in succession, etc. because we automatically refresh the dev config file whenever we save to it (causing new resource ids to get generated). To prevent this, we structure resource IDs as `devres{slug}`.

## Test plan
Tested various cases of deleting a resource.