Fix node entrypoint for bundle builds (#437)
## Description
The entrypoint should be the build version of the entrypoint which is always js. For non-inline tasks, since the definition is created by the user and could point to a ts file, we need to update the entrypoint to use the built js file.

Since inline task defs are parsed from the built js file, they dont have this problem.

## Test plan
Tested manually and ensured we can execute the task!