Remove feature that combines nested bundles (#447)
## Description
I added a feature that rolls together nested bundles. This was a good idea, but the backend bundler has as assumption baked into a few spots that the user code bundle it downloads from gcs is rooted at the task root. We could fix this assumption, but it's more trouble than it's worth, so I'm reverting this rollup change.

In it's place, I've added new logic (we discussed as part of the Projects prd discussion) that defaults the root (project base) of all tasks to the nearest airplane.yaml configuration file. This is a similar behavior as the rollup thing I removed in that it will ensure that nested tasks are included in the same project, but slightly different in that it doesn't work for nested projects.

https://media.giphy.com/media/pcJraL3NsSnqJdbyB2/giphy-downsized-large.gif

## Test plan
We previously had a bug where a nested task would build and deploy but fail with an error. I manually tested that it now works