Set env vars on task defn during discovery (#445)
## Description
Ensure that the task defn is the source of truth for env vars. Add in env vars from the airplane.yaml config file.

Without this change, we have to keep looking for and reading in the config file i.e. during local dev. With this change downstream code remains unchanged

## Test plan
Unit tested