Initialize an airplane config when creating Node.js tasks (#831)
## Description
Initialize an airplane.yaml config when creating a Node.js task. Eventually we're going to want to do this for all tasks, but starting with Node.js tasks.

I kept this simple. Originally I was trying to update the config and warn about mismatches. Updating the config isn't a great idea IMO as it will delete user comments and change the order of their file. Once they've created a config, leave them in charge of it.

I also tweaked some messaging.

## Test plan
Unit tests and manually tested this!