Always defualt views to the slim base (#473)
## Description
All views, inline or not, default to the slim base. Imoved this into a helper function rather than directly in the discoverer to airport can use the same helper function.

As written, the code ties the base of the view to the base of other JS tasks in the project as specified in the `airplane.yaml`. I **think** I like this as it gives users an out (they can use the full base for views) and if we do decide to build tasks and views together, they'll be of the same base.

In the future we could make a `views` section of airplane.yaml where you can specify the views base separately from your javascript tasks. I don't really think this is going to be a big deal though, and we could always do this in the future.

## Test plan
Tested manually e2e