Add isDir to file node (#852)
## Description

It is not enough to determine if a file node is a directory by checking if has any children—a directory may not have any files inside it. Instead, we attach an `isDir` field to each file node, which is preferable to checking if the path ends in `/`.

Also renames `App` -> `Entity`

## Test plan
Tested in conjunction with UI changes