Fix shell builds where base image has entrypoint (#482)
## Description
If the user supplied Dockerfile has an entrypoint, this entrypoint will be executed instead of our custom CMD/ARGS. We "unset" the entrypoint so it doesn't interfere.

I updated our testing infra a bit to reflect reality - we are using CMD not entrypoint. Without this I couldn't test.

## Test plan
I wrote a unit test that exactly replicated a user's issue. I then simplified it a bit.