Handle windows paths in discovery (#483)
## Description
This will happen on the clientside when we are discovering bundles and sending them to the server. On windows machines, we will set paths to `/`. The server applies these target paths against the user-uploaded code which is just a directory.

## Test plan
I will test this with our test windows machine after the fact. I don't have a great way to test locally on a windows machine :/

FIXES TSK-1304