Ignore GHA credentials (#864)
## Description
This change is a follow-up to https://github.com/airplanedev/cli/pull/799 to ignore the generated credentials file. The new steps in the GHA workflow are generating credentials files with a different name pattern, so we need to ignore the latter to avoid the "git is currently in a dirty state" errors from gorelreaser.

## Test plan
Testing to be completed by doing a new release after this is merged.