Populate default task params when executing local task (#865)
## Description
Populate default params in ExecuteTaskHandler. For task runs we rely on the ExecuteRunnableForm pre-filling default params, but this doesn't work for tasks called from views.

Fixes VIEW-613

## Test plan
Created a task with a default parameter, and a view with a task-backed button. Verified that task gets called with the default param, and the run is logged with the default param