Create views using inline config (#875)
## Description
Create views using inline config, behind a feature flag.

Inline views are not created in a sub-folder. I think we want to drop this behavior.

Inline views require an `airplane` dependency for doing the inline configuraiton.

## Test plan
Manually tested. It would be pretty valuable to set up unit tests for this. We'd have to find a way to mock out the filesystem.