Dynamically assign studio port (#870)
## Description

If a `--port` option is not provided, the studio will attempt to bind to port 4000. This is problematic if another process is already bound to this port, and the user will most likely run into this if they attempt to spin up multiple studio sessions. Instead, we choose a random open port.

## Test plan
Tested by spinning up two instances of the studio, and verified that the local dev server listened on two different ports. Also verified that the `--port` flag still works.