Support task in view file (#494)
## Description
Currently you can't define an inline task in a view because executing the task in a node environment fails with browser-specific concepts that are pulled in by the view.

I fixed this for discovery by replacing the document with JSDOM.

We have to do the same thing when builiding locally and remotely.

## Test plan
I added a unit test for a task defined in a view. It failed before and succeeds after this change.