Remove --production from npm/yarn install commands (#502)
## Description
We're getting a warning to stop using `--production` for JavaScript builds.

For both npm and yarn, this flag actually isn't needed since we are setting `NODE_ENV` to `production`.

```
With the --production flag (or when the NODE_ENV environment variable is set to production), npm will not install modules listed in devDependencies. 
```

## Test plan
All tests continue to pass