Support specifying task in a view file during local dev (#495)
## Description
See https://github.com/airplanedev/lib/pull/494 for context.

We are making the same change for local dev as we are making for deployment.

I took this one step further and made local dev use the new bundle based system so it is consistent with how the task will actually be deployed!

FIXES TSK-1361

## Test plan
I added a unit test for a task in a view