Esbuild remove css (#505)
## Description

FIXES PDS-1332

If a view imports css from a dependency, that css is marked as external meaning it is not bundled and is executed an runtime. When we execute the discovery parsing script, node attempts to import that css. Node does not know how to import css, so it errors.

In this pr, we are using an esbuild plugin to a) mark all css as non-external so it is bundled, and then when that bundled css is loaded, make it resolve to something that esbuild can load.

Inspiration from: https://github.com/evanw/esbuild/issues/619

## Test plan
Added unit tests