Add update files endpoint (#893)
## Description

Adds an endpoint `/dev/files/update` to the local dev server that will be used by the Studio to save files. For now, we just send over the contents of the file in the request, but in the future we could use SSE, WS, etc. to make this more efficient. Also adds a directory traversal check to `/dev/files/get`.

## Test plan
Tested in conjunction with web changes