Escape percent symbols when inlining into dockerfile (#509)
## Description

We use `printf` to inline into a dockerfile. % signs need to be escaped or else they are treated as format characters and an error is thrown.

We were seeing this if there was a `%` in the package.json. Now this is fixed.

FIXES PDS-1041

## Test plan
I wrote  unit test to cover this case. I also wrote an e2e test with `%` signs in the package.json that I was able to get passed, but I think this test is a bit contrived and I'm confident in the test of the helper function which is more generic and inclusive.