Incrementally find open studio port (#904)
## Description

Similar to Vite, we attempt to incrementally find a port for the local dev server starting at 4000, instead of getting a random open port. This is because users tend to refresh the studio page when restarting the studio, which is not possible if the port changes every time, and forces users to use the `--port` flag.

## Test plan

Tested `airplane dev` in two different tabs—the first dev server was bound to 4000 and the second one was bound to 4001. Also verified that the `--port` flag still worked.