Add endpoint to get view info (#909)
## Description

Implement endpoint to get view info in support of https://github.com/airplanedev/web/pull/2069.

## Test plan
I locally confirmed this works as expected.

I also confirmed that if the user is using an old version of the CLI, without this change but with https://github.com/airplanedev/web/pull/2069, the client behaves as expected.