Fix error checking (#911)
## Description

We were using `errors.As` wrong—the second argument should be a pointer to an empty error struct, _not_ and nil pointer. However, we don't use the resulting error anyway, and so we can perform a type assertion.

## Test plan
Verified that a view using a non-existent task slug properly shows `task with slug %s is not registered locally or remotely` instead of causing the CLI to panic.