Set view env vars from url params (#512)
## Description

We are now passing in view env vars as query params to the iframe in https://github.com/airplanedev/web/pull/2069

In this pr, we instruct the view shim (main.tsx) to read these query params and make them env vars that are accessible to the user's view as `process.env.TEAM_ID` e.g.

## Test plan
Manually tested. Here's `process.env` in a list:
![image](https://user-images.githubusercontent.com/6363419/213772398-c5ff71f2-3e19-4d8e-b8ba-5789bf96a14e.png)