Add endpoint to zip and return dev root (#910)
## Description

Adds an endpoint, `/dev/files/downloadBundle`, to the local dev server that zips the user's development workspace and returns it in the response. The archiving is done in-memory, but since we'll only be doing this for templates (and filtering out `node_modules/`, etc.), this should be ok.

## Test plan
Tested in conjunction with web changes.