Add optional auth middleware to dev server (#914)
## Description
We need to add auth middleware to the dev server for when the server is opened on a public ngrok tunnel. The middleware is optionally installed and checks for the presence of the correct token in the X-Airplane-Dev-Token header. I couldn't repurpose X-Airplane-Token since we use that to pass optional fake unverified JWTs to tasks.

## Test plan
Started studio with the --tunnel option and can't access the dev server publicly.