Update airplane ignore to handle yarn berry (#519)
## Description
Our patterns in our airplaneignore were incorrect for yarn berry. Docker ignore patterns are weird and what we had wasn't working. 

I updated so it works now.

## Test plan
I cloned a repo with yarn berry and confirmed that I could not deploy b/c the .yarn folder was not getting uploaded. After this change, I am able to deploy and the `releases` and `plugins` subfolders are uploaded. Other subfolders are not uploaded.

The test is commented back in in https://github.com/airplanedev/lib/pull/520