Add endpoint for dependency reinstallation + install dependencies on sandbox startup (#917)
## Description

This patch moves dependency installation to dev sandbox startup (instead of at the template cloning stage, which was the original plan). Partially this is because dependency installation can take a while (~10s for selenium), & it didn't make sense for it to be on the critical path for studio startup. This only happens in sandbox mode.

The `/dev/info` endpoint now also includes some sandbox-specific information, so the client can have some basic idea of the state of dependency management. Reinstallation of dependencies happens in an asynchronous mode because it's potentially a long operation; the `/dev/dependencies/reinstall` endpoint only triggers a reinstallation to happen.

Resolves DTS-112, DTS-165.

## Test plan

Tested locally.